Proper implementation of Decorator
