+++ src/main/java/pt/isec/mei/das/dto/BuildResultDTO.java
@@ -1,5 +1,7 @@
 package pt.isec.mei.das.dto;
 
+import java.math.BigDecimal;
+import java.math.RoundingMode;
 import java.time.LocalDateTime;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
