+++ src/main/java/pt/isec/mei/das/service/BuildService.java
@@ -181,7 +181,6 @@ public class BuildService {
   }
 
   public BuildResultDecoratedWithTimeInSeconds findDetailedBuildResultById(Long id) {
-
     BuildResult buildResult =
         buildResultRepository
             .findById(id)
